Cleanup / Fix native epoll multicast support
Motivation:

a3e8c867419b998c5da2174844b5f1e09ff76bed introduced multicast support (ipv4 only) when using native epoll transport but did not correctly handle errors in all cases and also did not choose the default InternetAddressFamily to use in the same way as we do in NioDatagramChannel. This could lead to silently dropping errors but also to testfailures if the loopback interface is configured with ipv4 and ipv6.

Beside this it also did not correctly support ipv6.

Modifications:

- Allow to specify the InternetAddressFamily when creating a EpollDatagramChannel (just as we do in NioDatagramChannel) and based on this select the correct address when configure an interface
- Throw exceptions if we can not init the sockaddr
- Add support for ipv6
- Add assume check to DatagramMulticastTest as it can only work if the Loopback interface also has ipv4 configured

Result:

Correctly handle epoll multicast with ipv4 and ipv6